chg: rename project from symphony to colusa
Colusa is not yet existed on pypi.org, so I rename the project
in order to be able to upload it to pypi.org

https://pypi.org/project/colusa
