+++ colusa/plugins/etr_cs_rutgers_edu.py
@@ -1,4 +1,4 @@
-from symphony.etr import Extractor, register_extractor
+from colusa.etr import Extractor, register_extractor
 
 
 @register_extractor('www.cs.rutgers.edu/~pxk/')
