+++ colusa/plugins/etr_medium.py
@@ -1,6 +1,6 @@
 from bs4 import Tag
 
-from symphony.etr import Extractor, Transformer, register_extractor, register_transformer
+from colusa.etr import Extractor, Transformer, register_extractor, register_transformer
 
 
 @register_extractor('//medium.com')
