+++ colusa/plugins/etr_truyenfull.py
@@ -2,7 +2,7 @@ import re
 
 from bs4 import Tag
 
-from symphony.etr import Extractor, Transformer, register_extractor, register_transformer
+from colusa.etr import Extractor, Transformer, register_extractor, register_transformer
 
 
 @register_extractor('//truyenfull.vn')
