+++ colusa/plugins/etr_unintendedconsequences.py
@@ -1,4 +1,4 @@
-from symphony.etr import Extractor, register_extractor
+from colusa.etr import Extractor, register_extractor
 
 
 @register_extractor('unintendedconsequenc')
